\docType{methods}
\name{collapseUnit}
\alias{collapseUnit}
\alias{collapseUnit,}
\alias{FLStock}
\title{Methods to collpase the units of an object}
\arguments{
  \item{object}{the thing you want to collapse}
}
\value{
  something the same class and dimension as \code{object}
  but with the unit dimension equal to 1
}
\description{
  currently implemented are collapsing accross units of an
  FLStock
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

