\name{frmtSeconds}
\alias{frmtSeconds}
\title{seconds to hrs, mins, secs}
\usage{
  frmtSeconds(x = 1)
}
\arguments{
  \item{x}{seconds}
}
\value{
  Some text
}
\description{
  seconds to hrs, mins, secs
}
\note{
  \code{helloWorld} is just a test
}
\examples{
helloWorld(3435)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

