\docType{methods}
\name{gislasim}
\alias{FLPar}
\alias{gislasim}
\alias{gislasim,}
\alias{numeric}
\title{Methods to generate lifehistory params}
\arguments{
  \item{linf}{the output from the mse function}
}
\value{
  an FLPar
}
\description{
  Methods to generate lifehistory params
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

