\name{helloWorld}
\alias{helloWorld}
\title{Hello world}
\usage{
  helloWorld(x = 1)
}
\arguments{
  \item{x}{random seed}
}
\value{
  Some text
}
\description{
  Hello world
}
\note{
  \code{helloWorld} is just a test
}
\examples{
helloWorld(3435)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

