\docType{methods}
\name{mseSumm}
\alias{list}
\alias{mseSumm}
\alias{mseSumm,}
\title{Methods to summaries Stock Simulations}
\arguments{
  \item{object}{the output from the mse function}
}
\value{
  a list
}
\description{
  Methods to summaries Stock Simulations
}
\seealso{
  \code{\link{print}} and \code{\link{cat}}
}

