\name{plotAssessment}
\alias{plotAssessment}
\title{Hello world}
\usage{
  plotAssessment(OM)
}
\arguments{
  \item{OM}{random seed}
}
\value{
  NULL
}
\description{
  Hello world
}
\note{
  \code{helloWorld} is just a test
}
\examples{
helloWorld(3435)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

