\name{plotOM}
\alias{plotOM}
\title{Hello world}
\usage{
  plotOM(OM, probs = c(0.025, 0.5, 0.975))
}
\arguments{
  \item{OM}{random seed}
}
\value{
  null
}
\description{
  Hello world
}
\note{
  \code{helloWorld} is just a test
}
\examples{
helloWorld(3435)
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}

