\docType{package}
\name{StockStructure-package}
\alias{StockStructure}
\alias{StockStructure-package}
\title{Functions to investige stock structure in fish}
\description{
  Functions to investige stock structure in fish
}
\details{
  \tabular{ll}{ Package: \tab StockStructure \cr Type: \tab
  Package \cr Version: \tab 0.1 \cr Date: \tab 2012-08-23
  \cr License: \tab GPL (>= 2) \cr LazyLoad: \tab yes \cr }
}
\author{
  Colin Millar \email{colin.millar@jrc.ec.europa.eu}

  Iago Mosqueira \email{iago.mosqueira@jrc.ec.europa.eu}
}
\references{
  none
}
\seealso{
  \code{\link{paste}}
}
\keyword{package}

